// src/schemas/SignupSchema.ts
import { z } from 'zod';

export const SignupSchema = z.object({
  name: z.string().min(1, { message: "Name is required" }),
  email: z.string().email({ message: "Invalid email address" }),
  password: z.string().min(6, { message: "Password must be at least 8 characters" }),
  role: z.enum(['admin', 'user']),
  // status: z.boolean(),
});

// Type inferred from the schema
export type ISignup = z.infer<typeof SignupSchema>;